Follow same pattern of .api method for RedirectQuerySet as well
